Migrate some tests that snuck into the main codebase to the test directory
