+++ src/tree.rs
@@ -55,7 +55,9 @@ impl TreeNode {
 /// Filter out Windows drive prefixes and root directory components to get logical path components.
 /// This ensures that paths like "C:\repo\src\lib.rs" become ["repo", "src", "lib.rs"]
 /// instead of ["C:", "\", "repo", "src", "lib.rs"].
-fn clean_path_components(path: &Path) -> Vec<String> {
+///
+/// Note: This function is public for testing purposes only.
+pub fn clean_path_components(path: &Path) -> Vec<String> {
     path.components()
         .filter_map(|component| match component {
             Component::Prefix(_) | Component::RootDir => None,
@@ -204,218 +206,5 @@ fn render_tree(node: &TreeNode, output: &mut String, prefix: &str, is_root: bool
     for (i, child) in children.iter().enumerate() {
         let is_last = i == children.len() - 1;
         render_child(child, output, prefix, is_last, is_root);
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use std::path::PathBuf;
-
-    #[test]
-    fn test_generate_tree_empty() {
-        let paths = vec![];
-        let result = generate_tree(&paths);
-        assert_eq!(result, "");
-    }
-
-    #[test]
-    fn test_generate_tree_single_file() {
-        let paths = vec![PathBuf::from("README.md")];
-        let result = generate_tree(&paths);
-        assert!(result.contains("Directory structure:"));
-        assert!(result.contains("└── README.md"));
-    }
-
-    #[test]
-    fn test_generate_tree_nested_structure() {
-        let paths = vec![
-            PathBuf::from("src/lib.rs"),
-            PathBuf::from("src/main.rs"),
-            PathBuf::from("Cargo.toml"),
-            PathBuf::from("README.md"),
-        ];
-        let result = generate_tree(&paths);
-
-        assert!(result.contains("Directory structure:"));
-        assert!(result.contains("├── src/"));
-        assert!(result.contains("│   ├── lib.rs"));
-        assert!(result.contains("│   └── main.rs"));
-        assert!(result.contains("├── Cargo.toml"));
-        assert!(result.contains("└── README.md"));
-    }
-
-    #[test]
-    fn test_generate_tree_directories_before_files() {
-        let paths = vec![PathBuf::from("file.txt"), PathBuf::from("dir/nested.rs")];
-        let result = generate_tree(&paths);
-
-        // Directories should come before files
-        let dir_pos = result.find("├── dir/").unwrap_or(0);
-        let file_pos = result.find("└── file.txt").unwrap_or(0);
-        assert!(dir_pos < file_pos);
-    }
-
-    #[test]
-    fn test_final_component_always_treated_as_file() {
-        // Test that final components are always treated as files, regardless of extension
-        let paths = vec![
-            PathBuf::from("Makefile"),      // No extension
-            PathBuf::from("Dockerfile"),    // No extension
-            PathBuf::from("src/mod"),       // No extension in subdirectory
-            PathBuf::from("config.toml"),   // With extension
-            PathBuf::from("scripts/build"), // No extension, could look like directory
-        ];
-        let result = generate_tree(&paths);
-
-        // All final components should be files (no trailing slash)
-        // Directories come first, then files alphabetically
-        assert!(result.contains("├── scripts/"));
-        assert!(result.contains("│   └── build")); // build should be a file, not build/
-        assert!(result.contains("├── src/"));
-        assert!(result.contains("│   └── mod")); // mod should be a file, not mod/
-        assert!(result.contains("├── Dockerfile"));
-        assert!(result.contains("├── Makefile"));
-        assert!(result.contains("└── config.toml")); // Last file uses └──
-
-        // Verify no final components have trailing slashes (which would indicate directories)
-        assert!(!result.contains("Dockerfile/"));
-        assert!(!result.contains("Makefile/"));
-        assert!(!result.contains("config.toml/"));
-        assert!(!result.contains("build/"));
-        assert!(!result.contains("mod/"));
-    }
-
-    #[test]
-    fn test_windows_path_component_filtering() {
-        // Test the clean_path_components function directly
-        // On Unix systems, we can't easily create Windows-style paths,
-        // so we test the filtering logic with relative paths that have
-        // problematic components like ".." and "."
-
-        let path = Path::new("./src/../src/lib.rs");
-        let components = clean_path_components(&path);
-
-        // Should filter out "." and keep ".." and normal components
-        assert_eq!(components, vec!["src", "..", "src", "lib.rs"]);
-
-        // Test with a simple path
-        let path = Path::new("repo/src/lib.rs");
-        let components = clean_path_components(&path);
-        assert_eq!(components, vec!["repo", "src", "lib.rs"]);
-    }
-
-    #[test]
-    fn test_path_normalization_in_tree() {
-        // Test that paths with current directory components are handled correctly
-        let paths = vec![PathBuf::from("./src/lib.rs"), PathBuf::from("src/main.rs")];
-        let result = generate_tree(&paths);
-
-        // Should contain proper structure without "./"
-        assert!(result.contains("└── src/"));
-        assert!(result.contains("    ├── lib.rs"));
-        assert!(result.contains("    └── main.rs"));
-        // Should not contain "./" in the output
-        assert!(!result.contains("./"));
-    }
-
-    #[test]
-    fn test_duplicate_file_paths() {
-        // Test that duplicate file paths are handled correctly
-        // The same file path added twice should still result in a single entry
-        let paths = vec![
-            PathBuf::from("src/lib.rs"),
-            PathBuf::from("src/lib.rs"), // Duplicate
-            PathBuf::from("src/main.rs"),
-        ];
-        let result = generate_tree(&paths);
-
-        // Should only show lib.rs once
-        let lib_rs_count = result.matches("lib.rs").count();
-        assert_eq!(
-            lib_rs_count, 1,
-            "lib.rs should appear only once, got: {}",
-            result
-        );
-
-        // Should still show both files
-        assert!(result.contains("├── lib.rs"));
-        assert!(result.contains("└── main.rs"));
-    }
-
-    #[test]
-    fn test_file_vs_directory_conflict() {
-        // Test when the same path is used as both intermediate directory and final file
-        // This tests the fix for issue where a file could be marked as directory
-        let paths = vec![
-            PathBuf::from("config/settings.json"), // config as directory
-            PathBuf::from("config"), // config as file - should be absorbed into directory
-            PathBuf::from("readme.txt"), // another file for comparison
-        ];
-        let result = generate_tree(&paths);
-
-        // config should be treated as a directory containing settings.json
-        // The standalone "config" file is absorbed into the directory structure
-        // because directory usage takes precedence when there are children
-        assert!(result.contains("├── config/"));
-        assert!(result.contains("│   └── settings.json"));
-        assert!(result.contains("└── readme.txt"));
-
-        // Should not show config as both file and directory
-        let config_lines: Vec<&str> = result
-            .lines()
-            .filter(|line| line.contains("config"))
-            .collect();
-        assert_eq!(
-            config_lines.len(),
-            1,
-            "Config should appear only once as directory"
-        );
-    }
-
-    #[test]
-    fn test_empty_directory_becomes_file() {
-        // Test that an empty directory entry can be converted to a file
-        let paths = vec![
-            PathBuf::from("item"), // item as file
-        ];
-        let result = generate_tree(&paths);
-
-        // item should be treated as a file (no trailing slash)
-        assert!(result.contains("└── item"));
-        assert!(!result.contains("item/"));
-    }
-
-    #[test]
-    fn test_processing_order_independence() {
-        // Test that the result is the same regardless of processing order
-        let paths1 = vec![
-            PathBuf::from("src/lib.rs"),
-            PathBuf::from("src/main.rs"),
-            PathBuf::from("src"),
-        ];
-        let paths2 = vec![
-            PathBuf::from("src"),
-            PathBuf::from("src/lib.rs"),
-            PathBuf::from("src/main.rs"),
-        ];
-
-        let result1 = generate_tree(&paths1);
-        let result2 = generate_tree(&paths2);
-
-        // Both should produce the same tree structure
-        // src should be a directory containing lib.rs and main.rs
-        assert!(result1.contains("src/"));
-        assert!(result1.contains("lib.rs"));
-        assert!(result1.contains("main.rs"));
-
-        assert!(result2.contains("src/"));
-        assert!(result2.contains("lib.rs"));
-        assert!(result2.contains("main.rs"));
-
-        // The essential structure should be the same (ignoring exact formatting)
-        let result1_lines: Vec<&str> = result1.lines().filter(|l| !l.trim().is_empty()).collect();
-        let result2_lines: Vec<&str> = result2.lines().filter(|l| !l.trim().is_empty()).collect();
-        assert_eq!(result1_lines.len(), result2_lines.len());
     }
 }
